search-engine-python/requirements.txt
sentence-transformers
faiss-cpu
rank-bm25
grpcio
grpcio-tools
numpy
ujson
symspellpy
tqdm
torch
